refactor: moved passwordIcon to components
